DOC-8942 add travel-sample docs
Mostly includes from sdk-common, as per others.

We could add an example / source listing to the middle section in future.
In the meantime I've just pointed at where to find the routes and the definitions.

{version-server} is the standard attribute name elsewhere (e.g. in includes)
so added this, though it's duplicated with cb_server_version

There's no swaggerui yet, so I've added the "readonly" API explorer link only.