CCBC-1320 Update Tracing and Metrics Docs
Filled the 4 pages, with samples.  Didn't actually put the sample
code in the repo, however, as it requires opentelemetry libraries to compile
and link.  I could add that to the CMakeList.txt, however we'd need
those libraries to be built as well (from source), which is a bit of
an effort.