DOC-8969 improve installation instructions
 * as "below" only makes sense on the Release page, and not the including pages,
   I've instead referred explicitly to the "Downloads table" which links to the
   Release Notes page #downloads (for the latest version)

 * Linked Windows/Mac/Linux words to their installation sections

 * Moved manual instructions up to top where they are already referenced.

 * Windows instructions: clarified "Windows" for the "Visual Studio" downloads,
   as this was the motivating issue for this ticket

 * Linux instructions:
   * Moved Redhat installation instructions out of the wrong section

   * Did a SIMPLE substitution of 3.2.0 for 3.2.1 in installation instructions,
     which had been missed from last release: will ask C devs to verify

      * (But did NOT update the "Verifying Installed Package" section as that includes
        checksums and similar which will also likely change)

   * renamed "Enterprise Linux" in download list to "Red Hat Enterprise Linux"

 * Mac instructions: added clarification that Homebrew only installs current version

 * Added note on Release page to update Installation instructions on release.
   (This could be changed to a partial include instead, if this doesn't help.)

"Windows binary packages can be found as downloads for each version listed below" only applies to the release notes page
even there, it's not entirely clear, as there isn't an obvious "list" (each version has a subheading and a lot of text before the list of binaries)
the text doesn't specify where below, so you have to scroll up and down to find it
(The "Verifying Installed Package" section is in between, and then there's a list of Jira tickets fixed in the release)
the "below" should probably link to #downloads at least as an example (e.g. to show the list of downloads for the most recent version)
The list of downloads doesn't repeat the phrase "Windows Binary" or anything about Windows...
(It mentions Visual Studio... which has a Mac version too, so this is ambiguous)
The sections on Linux are confusing: the heading "Configuring APT repositories (Debian, Ubuntu)" leads on to "For CentOS and Red Hat, the equivalent commands are:"
No binaries for Mac on the page.
Presumably only the latest version can be installed with Homebrew?
(There is wording about "To install development files and command line tools on Mac OS," but that doesn't mention "or versions other than the latest stable release")